Legacy shipping zone stock availability improvements (#19065)
* Adjust _get_best_gift_reward and add_variants_to_checkout

* Adjust validate_draft_order

* Adjust product filters

* Load site from data laoder in paylaod generation

* Improve the arguments for legacy behaviour

* Improve CheckoutLinesProblemsByCheckoutIdLoader

* Move add_variant_to_checkout to tests utils

* Adjust fulfillment webhooks

* Fix test

* Clear site cache in generate_checkout_payload

* Use site.get_current in _get_best_gift_reward

* Use site.get_current in get_available_warehouse_pks_for_product
