+++ saleor/checkout/tests/fixtures/checkout.py
@@ -8,7 +8,7 @@ from ....plugins.manager import get_plugins_manager
 from ....product.models import ProductVariantChannelListing
 from ...fetch import fetch_checkout_info, fetch_checkout_lines
 from ...models import Checkout, CheckoutDelivery, CheckoutLine, CheckoutMetadata
-from ...utils import add_variant_to_checkout
+from ..utils import add_variant_to_checkout
 
 
 @pytest.fixture
