+++ saleor/checkout/utils.py
@@ -7,7 +7,6 @@ from uuid import UUID
 
 import graphene
 from django.conf import settings
-from django.contrib.sites.models import Site
 from django.core.exceptions import ValidationError
 from django.db import transaction
 from django.db.models import prefetch_related_objects
@@ -19,7 +18,6 @@ from ..core.db.connection import allow_writer
 from ..core.exceptions import (
     NonExistingCheckout,
     NonExistingCheckoutLines,
-    ProductNotPublished,
 )
 from ..core.taxes import zero_taxed_money
 from ..core.utils.metadata_manager import (
@@ -56,7 +54,6 @@ from ..giftcard.utils import (
 from ..payment.models import Payment
 from ..plugins.manager import PluginsManager
 from ..product import models as product_models
-from ..warehouse.availability import check_stock_and_preorder_quantity
 from ..warehouse.reservations import reserve_stocks_and_preorders
 from . import AddressType, base_calculations, calculations
 from .delivery_context import is_shipping_required
@@ -175,121 +172,6 @@ def get_user_checkout(
     return checkout_queryset.filter(user=user, channel__is_active=True).first()
 
 
-def check_variant_in_stock(
-    checkout: Checkout,
-    variant: product_models.ProductVariant,
-    channel_slug: str,
-    quantity: int = 1,
-    replace: bool = False,
-    check_quantity: bool = True,
-    checkout_lines: list["CheckoutLine"] | None = None,
-    check_reservations: bool = False,
-) -> tuple[int, CheckoutLine | None]:
-    """Check if a given variant is in stock and return the new quantity + line."""
-    line = checkout.lines.filter(variant=variant).first()
-    line_quantity = 0 if line is None else line.quantity
-
-    new_quantity = quantity if replace else (quantity + line_quantity)
-
-    if new_quantity < 0:
-        raise ValueError(
-            f"{quantity!r} is not a valid quantity (results in {new_quantity!r})"
-        )
-
-    if new_quantity > 0 and check_quantity:
-        include_shipping_zones = Site.objects.get_current().settings.use_legacy_shipping_zone_stock_availability
-        check_stock_and_preorder_quantity(
-            variant,
-            checkout.get_country(),
-            channel_slug,
-            new_quantity,
-            include_shipping_zones=include_shipping_zones,
-            checkout_lines=checkout_lines,
-            check_reservations=check_reservations,
-        )
-
-    return new_quantity, line
-
-
-def add_variant_to_checkout(
-    checkout_info: "CheckoutInfo",
-    variant: product_models.ProductVariant,
-    quantity: int = 1,
-    price_override: Optional["Decimal"] = None,
-    replace: bool = False,
-    check_quantity: bool = True,
-    force_new_line: bool = False,
-):
-    """Add a product variant to checkout.
-
-    If `replace` is truthy then any previous quantity is discarded instead
-    of added to.
-
-    This function is not used outside of test suite.
-    """
-    checkout = checkout_info.checkout
-    channel_slug = checkout_info.channel.slug
-
-    product_channel_listing = product_models.ProductChannelListing.objects.filter(
-        channel_id=checkout.channel_id, product_id=variant.product_id
-    ).first()
-    if not product_channel_listing or not product_channel_listing.is_published:
-        raise ProductNotPublished()
-
-    variant_channel_listing = product_models.ProductVariantChannelListing.objects.get(
-        channel_id=checkout.channel_id, variant_id=variant.id
-    )
-    variant_price_amount = variant.get_base_price(
-        variant_channel_listing, price_override
-    ).amount
-    variant_prior_price_amount = variant.get_prior_price_amount(variant_channel_listing)
-
-    new_quantity, line = check_variant_in_stock(
-        checkout,
-        variant,
-        channel_slug,
-        quantity=quantity,
-        replace=replace,
-        check_quantity=check_quantity,
-    )
-
-    if force_new_line:
-        checkout.lines.create(
-            variant=variant,
-            quantity=quantity,
-            price_override=price_override,
-            undiscounted_unit_price_amount=variant_price_amount,
-            prior_unit_price_amount=variant_prior_price_amount,
-        )
-        return checkout
-
-    if line is None:
-        line = checkout.lines.filter(variant=variant).first()
-
-    if new_quantity == 0:
-        if line is not None:
-            line.delete()
-    elif line is None:
-        checkout.lines.create(
-            variant=variant,
-            quantity=new_quantity,
-            currency=checkout.currency,
-            price_override=price_override,
-            undiscounted_unit_price_amount=variant_price_amount,
-            prior_unit_price_amount=variant_prior_price_amount,
-        )
-    elif new_quantity > 0:
-        line.quantity = new_quantity
-        line.save(update_fields=["quantity"])
-
-    # invalidate calculated prices
-    price_expiration = timezone.now()
-    checkout.price_expiration = price_expiration
-    checkout.discount_expiration = price_expiration
-
-    return checkout
-
-
 def calculate_checkout_quantity(lines: list["CheckoutLineInfo"]):
     return sum([line_info.line.quantity for line_info in lines])
 
@@ -304,7 +186,7 @@ def add_variants_to_checkout(
     reservation_length: int | None = None,
     raise_error_for_missing_lines=False,
     *,
-    include_shipping_zones,
+    calculate_stocks_with_shipping_zones: bool,
 ):
     """Add variants to checkout.
 
@@ -401,7 +283,7 @@ def add_variants_to_checkout(
                 country_code,
                 channel,
                 reservation_length,
-                include_shipping_zones=include_shipping_zones,
+                calculate_stocks_with_shipping_zones=calculate_stocks_with_shipping_zones,
                 replace=replace_reservations,
             )
