+++ saleor/graphql/checkout/mutations/checkout_create.py
@@ -498,7 +498,7 @@ class CheckoutCreate(DeprecatedModelMutation, I18nMixin):
                     reservation_length=get_reservation_length(
                         site=site, user=info.context.user
                     ),
-                    include_shipping_zones=site.settings.use_legacy_shipping_zone_stock_availability,
+                    calculate_stocks_with_shipping_zones=site.settings.use_legacy_shipping_zone_stock_availability,
                 )
 
             # Save addresses
