+++ saleor/graphql/checkout/mutations/checkout_lines_add.py
@@ -134,7 +134,7 @@ class CheckoutLinesAdd(BaseMutation):
                         site=site, user=info.context.user
                     ),
                     raise_error_for_missing_lines=raise_error_for_missing_lines,
-                    include_shipping_zones=site.settings.use_legacy_shipping_zone_stock_availability,
+                    calculate_stocks_with_shipping_zones=site.settings.use_legacy_shipping_zone_stock_availability,
                 )
             except NonExistingCheckout as e:
                 graphql_id = graphene.Node.to_global_id("Checkout", e.checkout_token)
