+++ saleor/graphql/checkout/tests/mutations/test_checkout_add_promo_code.py
@@ -13,9 +13,7 @@ from .....checkout import base_calculations, calculations
 from .....checkout.actions import call_checkout_info_event
 from .....checkout.error_codes import CheckoutErrorCode
 from .....checkout.fetch import fetch_checkout_info, fetch_checkout_lines
-from .....checkout.utils import (
-    add_variant_to_checkout,
-)
+from .....checkout.tests.utils import add_variant_to_checkout
 from .....core.models import EventDelivery
 from .....discount import DiscountValueType, VoucherType
 from .....plugins.manager import get_plugins_manager
