+++ saleor/graphql/order/tests/test_draft_order_validate.py
@@ -18,6 +18,7 @@ def test_validate_draft_order(draft_order):
             draft_order.lines.all(),
             "US",
             get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
         is None
     )
@@ -33,6 +34,7 @@ def test_validate_draft_order_without_sku(draft_order):
             draft_order.lines.all(),
             "US",
             get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
         is None
     )
@@ -46,7 +48,11 @@ def test_validate_draft_order_wrong_shipping(draft_order):
     assert order.shipping_address.country.code not in shipping_zone.countries
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Shipping method is not valid for chosen shipping address"
     assert e.value.error_dict["shipping"][0].message == msg
@@ -56,7 +62,11 @@ def test_validate_draft_order_no_order_lines(order, shipping_method):
     order.shipping_method = shipping_method
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Could not create order without any products."
     assert e.value.error_dict["lines"][0].message == msg
@@ -72,7 +82,11 @@ def test_validate_draft_order_non_existing_variant(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Could not create orders with non-existing products."
     assert e.value.error_dict["lines"][0].message == msg
@@ -89,7 +103,11 @@ def test_validate_draft_order_with_unpublished_product(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Can't finalize draft with unpublished product."
     error = e.value.error_dict["lines"][0]
@@ -107,7 +125,11 @@ def test_validate_draft_order_with_unavailable_for_purchase_product(draft_order)
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Can't finalize draft with product unavailable for purchase."
     error = e.value.error_dict["lines"][0]
@@ -130,7 +152,11 @@ def test_validate_draft_order_with_product_available_for_purchase_in_future(
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Can't finalize draft with product unavailable for purchase."
     error = e.value.error_dict["lines"][0]
@@ -150,7 +176,11 @@ def test_validate_draft_order_out_of_stock_variant(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     msg = "Insufficient product stock."
     assert e.value.error_dict["lines"][0].message == msg
@@ -162,7 +192,11 @@ def test_validate_draft_order_no_shipping_address(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     error = e.value.error_dict["order"][0]
     assert error.message == "Can't finalize draft with no shipping address."
@@ -175,7 +209,11 @@ def test_validate_draft_order_no_billing_address(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     error = e.value.error_dict["order"][0]
     assert error.message == "Can't finalize draft with no billing address."
@@ -188,7 +226,11 @@ def test_validate_draft_order_no_shipping_method(draft_order):
 
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
     error = e.value.error_dict["shipping"][0]
     assert error.message == "Shipping method is required."
@@ -205,7 +247,11 @@ def test_validate_draft_order_no_shipping_method_shipping_not_required(
 
     assert (
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
         is None
     )
@@ -223,7 +269,11 @@ def test_validate_draft_order_no_shipping_address_no_method_shipping_not_require
 
     assert (
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
         is None
     )
@@ -237,7 +287,11 @@ def test_validate_draft_order_voucher(draft_order_with_voucher):
     # when & then
     with pytest.raises(ValidationError) as e:
         validate_draft_order(
-            order, order.lines.all(), "US", get_plugins_manager(allow_replica=False)
+            order,
+            order.lines.all(),
+            "US",
+            get_plugins_manager(allow_replica=False),
+            calculate_stocks_with_shipping_zones=True,
         ).get()
 
     error = e.value.error_dict["voucher"][0]
