+++ saleor/order/actions.py
@@ -583,8 +583,16 @@ def order_fulfilled(
             auto=auto,
         )
         webhook_events = [WebhookEventAsyncType.ORDER_UPDATED]
+        calculate_stocks_with_shipping_zones = (
+            site_settings.use_legacy_shipping_zone_stock_availability
+        )
         for fulfillment in fulfillments:
-            call_event(manager.fulfillment_created, fulfillment, notify_customer)
+            call_event(
+                manager.fulfillment_created,
+                fulfillment,
+                notify_customer,
+                calculate_stocks_with_shipping_zones,
+            )
 
         order_fulfilled = order.status == OrderStatus.FULFILLED
         if order_fulfilled:
@@ -592,7 +600,12 @@ def order_fulfilled(
 
         if order_fulfilled or manually_approved:
             for fulfillment in fulfillments:
-                call_event(manager.fulfillment_approved, fulfillment, notify_customer)
+                call_event(
+                    manager.fulfillment_approved,
+                    fulfillment,
+                    notify_customer,
+                    calculate_stocks_with_shipping_zones,
+                )
 
         call_order_events(
             manager,
@@ -796,6 +809,8 @@ def cancel_fulfillment(
     app: Optional["App"],
     warehouse: Optional["Warehouse"],
     manager: "PluginsManager",
+    *,
+    calculate_stocks_with_shipping_zones: bool,
 ):
     """Cancel fulfillment.
 
@@ -821,7 +836,11 @@ def cancel_fulfillment(
         fulfillment.status = FulfillmentStatus.CANCELED
         fulfillment.save(update_fields=["status"])
         update_order_status(fulfillment.order)
-        call_event(manager.fulfillment_canceled, fulfillment)
+        call_event(
+            manager.fulfillment_canceled,
+            fulfillment,
+            calculate_stocks_with_shipping_zones,
+        )
         call_order_event(
             manager,
             WebhookEventAsyncType.ORDER_UPDATED,
@@ -1076,7 +1095,7 @@ def _create_fulfillment_lines(
     gift_card_lines_info: list[GiftCardLineData],
     manager: "PluginsManager",
     *,
-    include_shipping_zones: bool,
+    calculate_stocks_with_shipping_zones: bool,
     should_decrease_stock: bool = True,
     allow_stock_to_be_exceeded: bool = False,
 ) -> list[FulfillmentLine]:
@@ -1101,8 +1120,8 @@ def _create_fulfillment_lines(
         should_decrease_stock (Bool): Stocks will get decreased if this is True.
         allow_stock_to_be_exceeded (bool): If `True` then stock quantity could exceed.
             Default value is set to `False`.
-        include_shipping_zones (bool): If `True`, stock is filtered by shipping
-            zones (legacy behavior). If `False`, all warehouse stocks in the
+        calculate_stocks_with_shipping_zones (bool): If `True`, stock is filtered by
+            shipping zones (legacy behavior). If `False`, all warehouse stocks in the
             channel are used regardless of shipping zones.
 
     Return:
@@ -1117,7 +1136,7 @@ def _create_fulfillment_lines(
     variants = [line.variant for line in lines if line.variant]
     stocks = (
         Stock.objects.for_channel_or_country(
-            channel_slug, include_shipping_zones=include_shipping_zones
+            channel_slug, include_shipping_zones=calculate_stocks_with_shipping_zones
         )
         .filter(warehouse_id=warehouse_pk, product_variant__in=variants)
         .select_related("product_variant")
@@ -1287,7 +1306,7 @@ def create_fulfillments(
                     manager,
                     should_decrease_stock=auto_approved,
                     allow_stock_to_be_exceeded=allow_stock_to_be_exceeded,
-                    include_shipping_zones=site_settings.use_legacy_shipping_zone_stock_availability,
+                    calculate_stocks_with_shipping_zones=site_settings.use_legacy_shipping_zone_stock_availability,
                 )
             )
             if tracking_number:
