+++ saleor/warehouse/reservations.py
@@ -34,7 +34,7 @@ def reserve_stocks_and_preorders(
     channel: "Channel",
     length_in_minutes: int,
     *,
-    include_shipping_zones: bool,
+    calculate_stocks_with_shipping_zones: bool,
     replace: bool = True,
 ):
     stock_variants, stock_lines = [], []
@@ -62,7 +62,7 @@ def reserve_stocks_and_preorders(
             channel,
             reserved_until,
             replace=replace,
-            include_shipping_zones=include_shipping_zones,
+            calculate_stocks_with_shipping_zones=calculate_stocks_with_shipping_zones,
         )
 
         # Refresh reserved_until for already existing lines
@@ -96,7 +96,7 @@ def reserve_stocks(
     reserved_until: datetime.datetime,
     *,
     replace: bool = True,
-    include_shipping_zones: bool,
+    calculate_stocks_with_shipping_zones: bool,
 ):
     """Reserve stocks for given `checkout_lines` in given country."""
     variants_ids = [line.variant_id for line in checkout_lines]
@@ -115,7 +115,7 @@ def reserve_stocks(
             channel.slug,
             variants,
             country_code=country_code,
-            include_shipping_zones=include_shipping_zones,
+            include_shipping_zones=calculate_stocks_with_shipping_zones,
         )
         .order_by("pk")
         .values("id", "product_variant", "pk", "quantity", "warehouse_id")
