+++ saleor/warehouse/tests/test_stock_reservations_management.py
@@ -27,7 +27,7 @@ def test_reserve_stocks(checkout_line, channel_USD):
         COUNTRY_CODE,
         channel_USD,
         timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-        include_shipping_zones=True,
+        calculate_stocks_with_shipping_zones=True,
     )
 
     stock.refresh_from_db()
@@ -48,7 +48,7 @@ def test_stocks_reservation_skips_prev_reservation_delete_if_replace_is_disabled
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
             replace=False,
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
     with assert_num_queries(4):
@@ -58,7 +58,7 @@ def test_stocks_reservation_skips_prev_reservation_delete_if_replace_is_disabled
             COUNTRY_CODE,
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -92,7 +92,7 @@ def test_multiple_stocks_reserved_if_single_stock_is_not_enough_highest_stock_st
         COUNTRY_CODE,
         channel_USD,
         timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-        include_shipping_zones=True,
+        calculate_stocks_with_shipping_zones=True,
     )
 
     stock.refresh_from_db()
@@ -162,7 +162,7 @@ def test_multiple_stocks_reserved_if_single_stock_is_not_enough_sorting_order_st
         COUNTRY_CODE,
         channel_USD,
         timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-        include_shipping_zones=True,
+        calculate_stocks_with_shipping_zones=True,
     )
 
     # then
@@ -205,7 +205,7 @@ def test_stocks_reservation_removes_previous_reservations_for_checkout(
         COUNTRY_CODE,
         channel_USD,
         timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-        include_shipping_zones=True,
+        calculate_stocks_with_shipping_zones=True,
     )
 
     with pytest.raises(Reservation.DoesNotExist):
@@ -229,7 +229,7 @@ def test_stock_reservation_fails_if_there_is_not_enough_stock_available(
             COUNTRY_CODE,
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -246,7 +246,7 @@ def test_stock_reservation_fails_if_there_is_no_stock(checkout_line, channel_USD
             COUNTRY_CODE,
             channel_USD,
             RESERVATION_LENGTH,
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -271,7 +271,7 @@ def test_stock_reservation_accounts_for_order_allocations(
             COUNTRY_CODE,
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -314,7 +314,7 @@ def test_stock_reservation_accounts_for_order_allocations_and_reservations(
             COUNTRY_CODE,
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -337,7 +337,7 @@ def test_reserve_stocks_no_shipping_zones_included(checkout_line, channel_USD):
             COUNTRY_CODE,
             channel_USD,
             timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-            include_shipping_zones=True,
+            calculate_stocks_with_shipping_zones=True,
         )
 
 
@@ -361,7 +361,7 @@ def test_reserve_stocks_no_shipping_zones_excluded_from_stock_calculations(
         COUNTRY_CODE,
         channel_USD,
         timezone.now() + datetime.timedelta(minutes=RESERVATION_LENGTH),
-        include_shipping_zones=False,
+        calculate_stocks_with_shipping_zones=False,
     )
 
     # then - reservation created
