+++ saleor/webhook/payloads.py
@@ -552,10 +552,14 @@ def generate_checkout_payload(
 
     # todo use the most appropriate warehouse
     warehouse = None
-    include_shipping_zones = (
-        Site.objects.get_current().settings.use_legacy_shipping_zone_stock_availability
+
+    # refetch the site to make sure that we have the latest settings
+    Site.objects.clear_cache()
+    site = Site.objects.get_current()
+    calculate_stocks_with_shipping_zones = (
+        site.settings.use_legacy_shipping_zone_stock_availability
     )
-    if include_shipping_zones:
+    if calculate_stocks_with_shipping_zones:
         if checkout.shipping_address:
             warehouse = Warehouse.objects.for_country_and_channel(
                 checkout.shipping_address.country.code, checkout.channel_id
@@ -985,7 +989,9 @@ def generate_fulfillment_lines_payload(fulfillment: Fulfillment):
 @allow_writer()
 @traced_payload_generator
 def generate_fulfillment_payload(
-    fulfillment: Fulfillment, requestor: Optional["RequestorOrLazyObject"] = None
+    fulfillment: Fulfillment,
+    requestor: Optional["RequestorOrLazyObject"] = None,
+    calculate_stocks_with_shipping_zones: bool = True,
 ):
     serializer = PayloadSerializer()
 
@@ -1010,8 +1016,7 @@ def generate_fulfillment_payload(
     if fulfillment_line and fulfillment_line.stock:
         warehouse = fulfillment_line.stock.warehouse
     else:
-        include_shipping_zones = Site.objects.get_current().settings.use_legacy_shipping_zone_stock_availability
-        if include_shipping_zones:
+        if calculate_stocks_with_shipping_zones:
             warehouse = Warehouse.objects.for_country_and_channel(
                 order_country, order.channel_id
             ).first()
